requirements.txt
#---------------------------------------------------------
# py -m pip install -r requirements.txt
#---------------------------------------------------------
requests
datetime
ruamel.yaml
tqdm
notion-python-2